Namespaced the `Faraday` middleware classes to fix conflicts with the `twitter` gem.
